Initial working version of main page ad carousel-slider
Adds a new <promotergallery /> tag, which then tries to load
all ads for the current page.

This uses 2 new modules: jquery.owl-carousel and an initiator for it
There is also a new SpecialCampaignAdsLoader, used to get all ads
for a campaign instead of just a single random one - but currently it
is not used and may be removed at a later time.